+++ base/README.md
@@ -1,1 +1,25 @@
-# BASE SDK
+![base-sdk](../_branding/cover.png)
+
+# BASE SDK base
+
+## Installation
+
+```sh
+yarn add @base-sdk/base
+```
+
+## Usage
+
+<!-- add sdk usage example here -->
+
+## What is BASE?
+
+Base is an opensource firebase alternative for building enterprise level applications. the backend runs on lambda as MSA, and you can self-host it if you want. the backend is available at [bridgedxyz/base](https://github.com/bridgedxyz/base)
+
+## Follow us & Learn more
+
+-   [bridged homepage](https://bridged.xyz)
+-   [bridged github](https://github.com/bridgedxyz)
+-   [base backend](https://github.com/bridgedxyz/base)
+-   [bridged on facebook](https://www.facebook.com/bridged.xyz/)
+-   [bridged on twitter](https://twitter.com/bridgedxyz)
