+++ base/index.ts
@@ -2,16 +2,16 @@
 export const PACKAGE_NAME = "@base-sdk/base";
 
 // exports
+export * as url from "@base-sdk/url";
+export * as hosting from "@base-sdk/url";
+export * as cors from "@base-sdk/core/cors";
 export * as _abstract from "./services-abstract";
 export * as projects from "./projects";
 export * as features from "./features";
 export * as assets from "./assets";
 export * from "./asset-repository";
-export * as url from "./url";
-export * as cors from "./cors";
 export * as resources from "./resources";
 export * from "./design";
 export * as versions from "./versions";
 export * from "./scene";
-export * as hosting from "./hosting";
 export * as types from "./types";
