+++ core/_utils/index.ts
@@ -1,0 +1,7 @@
+///
+/// exsport pattern
+/// `export * as url from "./url-utils";`
+/// will be accessable as `_utils.url.*`
+///
+
+export * as url from "./url-utils";
