+++ core/index.ts
@@ -3,3 +3,8 @@ export const PACKAGE_NAME = "@base-sdk/core";
 
 // exports internal constants
 export * as __HOSTS from "./constants/hosts";
+export * as cors from "./cors";
+
+// export _utils for internal usage
+export * as _utils from "./_utils";
+export * as utils_url from "./_utils/url-utils";
