--- docs/README.md
@@ -1,1 +1,0 @@
-# base-sdk for nodejs Docs directory
