+++ hosting/index.ts
@@ -1,0 +1,5 @@
+// preserve
+export const PACKAGE_NAME = "@base-sdk/hosting";
+
+export * from "./types";
+export * from "./upload";
