+++ hosting/types/index.ts
