+++ url/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../tsconfig.json",
+    "compilerOptions": {
+        "outDir": "./dist",
+        "baseUrl": "./"
+    },
+    "exclude": ["node_modules", "example", "dist", "docs"]
+}
